requirements.txt
python-telegram-bot
mutagen
Pillow
pyrogram
tgcrypto
yt-dlp
scdl
httpx
apscheduler
python-dotenv